Consider only public instance properties in DeltaOfEntityType
Using JSON.NET serializer in a non-EntityFramework scenario: When the entities contain static properties, trying to PATCH with Delta fails with an invalid model state (Cannot bind to the target method because its signature or security transparency is not compatible with that of the delegate type). Retrieving only instance properties fixes the error.